refactor: refactor action log
