+++ packages/client/src/action-logs/ActionLog.tsx
@@ -1,10 +1,32 @@
 import { Field } from '@formily/core';
 import { observer, RecursionField, useField, useFieldSchema } from '@formily/react';
-import React from 'react';
+import React, { useContext } from 'react';
 import { useCollection } from '..';
+import { SchemaComponent } from '../schema-component';
+import { SchemaInitializerContext } from '../schema-initializer';
 import { ActionLogDesigner } from './ActionLog.Designer';
+import { createSchema } from './utils';
 
-export const ActionLog = () => null;
+export const ActionLog = () => {
+  const initializers = useContext(SchemaInitializerContext);
+  initializers.BlockInitializers.items.push({
+    type: 'itemGroup',
+    title: '{{t("others")}}',
+    children: [
+      {
+        type: 'item',
+        title: '{{t("operation log")}}',
+        component: 'ActionLogBlockInitializer',
+      },
+    ],
+  });
+  const schema = createSchema();
+  return (
+    <SchemaInitializerContext.Provider value={initializers}>
+      <SchemaComponent schema={schema} name={schema.name} />
+    </SchemaInitializerContext.Provider>
+  );
+};
 
 ActionLog.Designer = ActionLogDesigner;
