+++ packages/client/src/locale/en_US.json
@@ -277,5 +277,7 @@
   "Default": "Default",
   "Add card": "Add card",
   "edit title": "Edit Title",
-  "turn page": "Turn Page"
+  "turn page": "Turn Page",
+  "others": "others",
+  "operation log": "operation log"
 }
