+++ packages/client/src/locale/zh_CN.json
@@ -279,5 +279,7 @@
 
   "Add card": "添加卡片",
   "edit title": "修改标题",
-  "turn page": "翻页"
+  "turn page": "翻页",
+  "others": "其他",
+  "operation log": "操作记录"
 }
