+++ packages/client/src/schema-initializer/Initializers/BlockInitializers.tsx
@@ -42,16 +42,5 @@ export const BlockInitializers = {
         },
       ],
     },
-    {
-      type: 'itemGroup',
-      title: 'Templates',
-      children: [
-        {
-          type: 'item',
-          title: 'Action logs',
-          component: 'ActionLogBlockInitializer',
-        },
-      ],
-    },
   ],
 };
